tests/agents/test_frontend_agent.py
"""Tests for the frontend agent."""

import pytest
from agents.specialized.frontend import FrontendAgent
from core.config_models import FrontendAgentConfig

@pytest.fixture
def frontend_agent():
    """Create a frontend agent for testing."""
    config = FrontendAgentConfig(
        agent_id="test_frontend",
        component_dir="tests/data/components",
        default_framework="react",
        supported_frameworks=["react", "vue"],
        build_throttle=1.0,
        requires_approval=False
    )
    return FrontendAgent("test_frontend", config)

def test_frontend_agent_initialization(frontend_agent):
    """Test frontend agent initialization."""
    assert frontend_agent.agent_id == "test_frontend"
    assert frontend_agent.config.default_framework == "react"
    assert "react" in frontend_agent.config.supported_frameworks

def test_frontend_agent_component_creation(frontend_agent):
    """Test component creation."""
    # Add component creation tests here
    pass

def test_frontend_agent_validation(frontend_agent):
    """Test validation rules."""
    # Add validation tests here
    pass

def test_frontend_agent_throttling(frontend_agent):
    """Test build throttling."""
    # Add throttling tests here
    pass 