requirements.txt
pyyaml>=6.0.0
rich>=13.0.0
openai>=1.0.0
python-dotenv>=1.0.0
loguru>=0.7.0
pydantic>=2.0.0
chromadb>=0.4.0
streamlit>=1.24.0
pandas>=2.0.0
docker>=6.0.0
fastapi>=0.100.0
uvicorn>=0.22.0
aiohttp>=3.8.0
requests>=2.31.0
cryptography>=41.0.0
PyPDF2>=3.0.0
python-docx>=0.8.11
openpyxl>=3.1.0
beautifulsoup4>=4.12.0
SpeechRecognition>=3.10.0
PyGithub>=2.1.1
watchdog>=3.0.0 